getUnique

getUnique

Usage

getUnique(.data)

Arguments

.data

input data as used in a pipe

Value

Dataframe where the rows are the columns of .data. For each variable the function returns the number of unique values, nu_unique, and what these values are, unique_values. NOTE: Columns that are constant are not shown.

Examples

data(mtcars)
mtcars %>% getUnique()
#> # A tibble: 11 × 3
#>    column nu_unique unique_values                                               
#>    <chr>      <dbl> <chr>                                                       
#>  1 qsec          30 16.46,17.02,18.61,19.44,20.22,15.84,20,22.9,18.3,18.9,17.4,…
#>  2 wt            29 2.62,2.875,2.32,3.215,3.44,3.46,3.57,3.19,3.15,4.07,3.73,3.…
#>  3 disp          27 160,108,258,360,225,146.7,140.8,167.6,275.8,472,460,440,78.…
#>  4 mpg           25 21,22.8,21.4,18.7,18.1,14.3,24.4,19.2,17.8,16.4,17.3,15.2,1…
#>  5 hp            22 110,93,175,105,245,62,95,123,180,205,215,230,66,52,65,97,15…
#>  6 drat          22 3.9,3.85,3.08,3.15,2.76,3.21,3.69,3.92,3.07,2.93,3,3.23,4.0…
#>  7 carb           6 4,1,2,3,6,8                                                 
#>  8 cyl            3 6,4,8                                                       
#>  9 gear           3 4,3,5                                                       
#> 10 vs             2 0,1                                                         
#> 11 am             2 1,0